// src/plugins/ws/protobuf/ts/proto/PerpetualGroupDetail.ts
export interface PerpetualGroupDetail {
    /**
     * @generated from protobuf field: PerpetualGroupDetail.Detail detail = 1;
     */
    detail?: PerpetualGroupDetail_Detail; // 合约组详情
    /**
     * @generated from protobuf field: repeated PerpetualGroupDetail.Positions positions = 2;
     */
    positions: PerpetualGroupDetail_Positions[]; // 合约组持仓列表
}
/**
 * @generated from protobuf message PerpetualGroupDetail.Detail
 */
export interface PerpetualGroupDetail_Detail {
    /**
     * @generated from protobuf field: string groupAsset = 1;
     */
    groupAsset: string; // 合约组总价值
    /**
     * @generated from protobuf field: string marginAvailable = 2;
     */
    marginAvailable: string; // 合约组可用保证金
    /**
     * @generated from protobuf field: string positionMargin = 3;
     */
    positionMargin: string; // 仓位保证金
    /**
     * @generated from protobuf field: repeated PerpetualGroupDetail.MarginCoin marginCoin = 4;
     */
    marginCoin: PerpetualGroupDetail_MarginCoin[]; // 保证金占比
    /**
     * @generated from protobuf field: repeated PerpetualGroupDetail.PositionAsset positionAsset = 5;
     */
    positionAsset: PerpetualGroupDetail_PositionAsset[]; // 持仓风险占比
    /**
     * @generated from protobuf field: string groupId = 6;
     */
    groupId: string; // 合约组 ID
    /**
     * @generated from protobuf field: string baseCoin = 7;
     */
    baseCoin: string; // 计价币
    /**
     * @generated from protobuf field: string groupName = 8;
     */
    groupName: string; // 合约组名称
    /**
     * @generated from protobuf field: string availableBalanceValue = 9;
     */
    availableBalanceValue: string; // 现货资产购买力
}
/**
 * @generated from protobuf message PerpetualGroupDetail.MarginCoin
 */
export interface PerpetualGroupDetail_MarginCoin {
    /**
     * @generated from protobuf field: string coinName = 1;
     */
    coinName: string; // 币种名称
    /**
     * @generated from protobuf field: string coinConvert = 2;
     */
    coinConvert: string; // 折算价值
}
/**
 * @generated from protobuf message PerpetualGroupDetail.PositionAsset
 */
export interface PerpetualGroupDetail_PositionAsset {
    /**
     * @generated from protobuf field: string groupId = 1;
     */
    groupId: string; // 合约组 ID
    /**
     * @generated from protobuf field: string groupName = 2;
     */
    groupName: string; // 合约组名称
    /**
     * @generated from protobuf field: string baseCoin = 3;
     */
    baseCoin: string; // 计价币
}
/**
 * @generated from protobuf message PerpetualGroupDetail.Positions
 */
export interface PerpetualGroupDetail_Positions {
    /**
     * @generated from protobuf field: string groupId = 1;
     */
    groupId: string; // 合约组 id
    /**
     * @generated from protobuf field: string groupName = 2;
     */
    groupName: string; // 合约组名称
    /**
     * @generated from protobuf field: string groupMargin = 3;
     */
    groupMargin: string; // 合约组保证金
    /**
     * @generated from protobuf field: string tradeId = 4;
     */
    tradeId: string; // 交易对 id
    /**
     * @generated from protobuf field: string symbol = 5;
     */
    symbol: string; // 交易对名称
    /**
     * @generated from protobuf field: string baseSymbolName = 6;
     */
    baseSymbolName: string; // 基础币
    /**
     * @generated from protobuf field: string quoteSymbolName = 7;
     */
    quoteSymbolName: string; // 计价币
    /**
     * @generated from protobuf field: string typeInd = 8;
     */
    typeInd: string; // 合约类型
    /**
     * @generated from protobuf field: string positionId = 9;
     */
    positionId: string; // 仓位 id
    /**
     * @generated from protobuf field: string sideInd = 10;
     */
    sideInd: string; // long:多仓; short:空仓
    /**
     * @generated from protobuf field: string lever = 11;
     */
    lever: string; // 杠杆倍数
    /**
     * @generated from protobuf field: string unrealizedProfit = 12;
     */
    unrealizedProfit: string; // 未实现盈亏
    /**
     * @generated from protobuf field: string profitRatio = 13;
     */
    profitRatio: string; // 收益率
    /**
     * @generated from protobuf field: string size = 14;
     */
    size: string; // 持仓数量
    /**
     * @generated from protobuf field: string marginRatio = 15;
     */
    marginRatio: string; // 仓位保证金率
    /**
     * @generated from protobuf field: string maintMarginRatio = 16;
     */
    maintMarginRatio: string; // 维持保证金率
    /**
     * @generated from protobuf field: string openPrice = 17;
     */
    openPrice: string; // 开仓均价
    /**
     * @generated from protobuf field: string markPrice = 18;
     */
    markPrice: string; // 标记价格
    /**
     * @generated from protobuf field: string latestPrice = 19;
     */
    latestPrice: string; // 最新价
    /**
     * @generated from protobuf field: string liquidatePrice = 20;
     */
    liquidatePrice: string; // 预估强平价
    /**
     * @generated from protobuf field: string realizedProfit = 21;
     */
    realizedProfit: string; // 已实现盈亏
    /**
     * @generated from protobuf field: string profit = 22;
     */
    profit: string; // 收益
    /**
     * @generated from protobuf field: string statusCd = 23;
     */
    statusCd: string; // 仓位状态：opened:已开仓，closed:已关闭，locked:锁仓中
    /**
     * @generated from protobuf field: string lockSize = 24;
     */
    lockSize: string; // 锁仓数量
    /**
     * @generated from protobuf field: string lockFees = 25;
     */
    lockFees: string; // 锁仓利息
    /**
     * @generated from protobuf field: string amountOffset = 26;
     */
    amountOffset: string; // 数量精度
    /**
     * @generated from protobuf field: string priceOffset = 27;
     */
    priceOffset: string; // 价格精度
    /**
     * @generated from protobuf field: string createdByTime = 28;
     */
    createdByTime: string; // 创建时间
    /**
     * @generated from protobuf field: string updatedByTime = 29;
     */
    updatedByTime: string; // 修改时间
    /**
     * @generated from protobuf field: PerpetualGroupDetail.LockRecord lockRecord = 30;
     */
    lockRecord?: PerpetualGroupDetail_LockRecord; // 锁仓记录
    /**
     * @generated from protobuf field: repeated PerpetualGroupDetail.ProfitLossEntrust profitLossEntrust = 31;
     */
    profitLossEntrust: PerpetualGroupDetail_ProfitLossEntrust[]; // 仓位止盈止损
    /**
     * @generated from protobuf field: string entrustFrozenSize = 32;
     */
    entrustFrozenSize: string; // 委托冻结的数量
}
/**
 * @generated from protobuf message PerpetualGroupDetail.LockRecord
 */
export interface PerpetualGroupDetail_LockRecord {
    /**
     * @generated from protobuf field: string lockPrice = 1;
     */
    lockPrice: string; // 首次锁仓价格
    /**
     * @generated from protobuf field: string fees = 2;
     */
    fees: string; // 锁仓手续费总和
}
/**
 * @generated from protobuf message PerpetualGroupDetail.ProfitLossEntrust
 */
export interface PerpetualGroupDetail_ProfitLossEntrust {
    /**
     * @generated from protobuf field: string strategyTypeInd = 1;
     */
    strategyTypeInd: string; // 策略类型 stop_profit 止盈 stop_loss 止损
    /**
     * @generated from protobuf field: string triggerPrice = 2;
     */
    triggerPrice: string; // 触发价格
    /**
     * @generated from protobuf field: string triggerPriceTypeInd = 3;
     */
    triggerPriceTypeInd: string; // 触发方式：mark 标记，new 最新
    /**
     * @generated from protobuf field: string id = 4;
     */
    id: string; // 仓位止盈止损 id
}
// @generated message type with reflection information, may provide speed optimized methods
class PerpetualGroupDetail$Type extends MessageType<PerpetualGroupDetail> {
    constructor() {
        super("PerpetualGroupDetail", [
            { no: 1, name: "detail", kind: "message", T: () => PerpetualGroupDetail_Detail },
            { no: 2, name: "positions", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => PerpetualGroupDetail_Positions }
        ]);
    }
    create(value?: PartialMessage<PerpetualGroupDetail>): PerpetualGroupDetail {
        const message = { positions: [] };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<PerpetualGroupDetail>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PerpetualGroupDetail): PerpetualGroupDetail {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* PerpetualGroupDetail.Detail detail */ 1:
                    message.detail = PerpetualGroupDetail_Detail.internalBinaryRead(reader, reader.uint32(), options, message.detail);
                    break;
                case /* repeated PerpetualGroupDetail.Positions positions */ 2:
                    message.positions.push(PerpetualGroupDetail_Positions.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PerpetualGroupDetail, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* PerpetualGroupDetail.Detail detail = 1; */
        if (message.detail)
            PerpetualGroupDetail_Detail.internalBinaryWrite(message.detail, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated PerpetualGroupDetail.Positions positions = 2; */
        for (let i = 0; i < message.positions.length; i++)
            PerpetualGroupDetail_Positions.internalBinaryWrite(message.positions[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message PerpetualGroupDetail
 */
export const PerpetualGroupDetail = new PerpetualGroupDetail$Type();
// ...
export const PerpetualGroupDetail_Detail = new PerpetualGroupDetail_Detail$Type();
// ...
export const PerpetualGroupDetail_MarginCoin = new PerpetualGroupDetail_MarginCoin$Type();
// ...
export const PerpetualGroupDetail_PositionAsset = new PerpetualGroupDetail_PositionAsset$Type();
// ...
export const PerpetualGroupDetail_Positions = new PerpetualGroupDetail_Positions$Type();
// ...
export const PerpetualGroupDetail_LockRecord = new PerpetualGroupDetail_LockRecord$Type();
// ...
export const PerpetualGroupDetail_ProfitLossEntrust = new PerpetualGroupDetail_ProfitLossEntrust$Type();